Phase 16: clear remaining SonarCloud smells (complexity, init-statement, casts, nested break)
Fixed 7 SonarCloud quality issues:
1. audio_decoder.cpp:126 (S1905) - removed redundant cast to int
2. audio_decoder.cpp:127-128 (S6004) - moved fmt into interleave_to_f32 call
3. audio_decoder.cpp:99 (S924) - reduced nested break from 2 to 1 via while condition
4. audio_interleave.cpp:44 (S5276) - added explicit cast of int32_t sample to float
5. video_decoder.cpp:284 (S6004) - init-statement: moved ret into if condition
6. video_decoder.cpp:384 (S3776) - reduced complexity by extracting handle_eagain_case helper
7. video_decoder.cpp:462 (S924) - reduced nested break from 2 to 1 by inverting loop condition

All changes are quality-only; behavior is identical. All 411 tests pass with ASAN.

+++ src/media/audio_interleave.cpp
@@ -41,7 +41,7 @@ inline float read_sample_as_float<int32_t>(const uint8_t* plane, int index)
     int32_t sample = 0;
     std::memcpy(&sample, plane + index * static_cast<int>(sizeof(int32_t)),
                 sizeof(int32_t));
-    return sample / 2147483648.0f;
+    return static_cast<float>(sample) / 2147483648.0f;
 }
 
 template <>
